refactor posts and replies actions folder
