+++ frontend/src/actions/index.ts
@@ -1,3 +1,4 @@
 export * from "./projects.server";
-export * from "./projectPosts.server";
+export * from "./project-posts.server";
+export * from "./users.server";
 export * from "./auth.server";
