+++ frontend/src/actions/project-posts.server/createDislikePost.server.ts
@@ -1,0 +1,34 @@
+"use server";
+
+import { auth } from "@/auth";
+import { baseUrl } from "@/constants";
+import { ProjectPostId } from "@/types/ProjectPost";
+
+// POST Dislike a post
+const createDislikePost = async (projectPostId: ProjectPostId) => {
+  const session = await auth();
+  try {
+    const response = await fetch(
+      `${baseUrl}/api/project-posts/${projectPostId}/dislike`,
+      {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({
+          userId: session?.user?.id,
+        }),
+      },
+    );
+    if (!response.ok) {
+      throw new Error(`Failed to dislike post: ${response.statusText}`);
+    }
+    const data = await response.json();
+    console.log("Disliked data:", data);
+    return data.dislikes;
+  } catch (error) {
+    console.error("Error disliking post:", error);
+  }
+};
+
+export { createDislikePost };
