+++ frontend/src/actions/users.server/getUserByUsername.server.ts
@@ -1,9 +1,9 @@
+"use server";
 import { auth } from "@/auth";
 import { baseUrl } from "@/constants";
 import type { User } from "@/types/User";
 
 const getUserByUsername = async () => {
-  "use server";
   const session = await auth();
   console.log(session);
   if (!session || !session.user || !session.user.githubUsername) {
@@ -22,4 +22,4 @@ const getUserByUsername = async () => {
   }
 };
 
-export default getUserByUsername;
+export { getUserByUsername };
